* Go And Spin Enemy 패턴 추가 * Back And Forth Enemy 패턴 추가 * Stage 내 Enemy Spawn Script 추가

+++ ShootingStrike/TextUI.cpp
@@ -36,7 +36,7 @@ void TextUI::Render(HDC _hdc)
 
 	if ( pTextImage && isValidText )
 	{
-		for ( int textOrder = 0; textOrder < splitTextList.size(); ++textOrder )
+		for ( int textOrder = 0; textOrder < (int)splitTextList.size(); ++textOrder )
 		{			
 			eTextType textType = splitTextList[textOrder].first;
 			int segmentOffset = splitTextList[textOrder].second;
